move sync to separate package
